backend/requirements.txt
agno
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
faker
sqlalchemy
psycopg2-binary
psycopg2